+++ lizard_damage/management/commands/clean_up.py
@@ -8,7 +8,7 @@ from __future__ import (
 
 from django.core.management.base import BaseCommand
 
-from lizard_damage import models
+from lizard_damage.models import DamageScenario
 from lizard_task.models import SecuredPeriodicTask
 
 import logging
@@ -24,8 +24,8 @@ class Command(BaseCommand):
     def handle(self, *args, **options):
         logger.info("Cleaning up scenarios which are expired...")
         now = datetime.datetime.now()
-        for damage_scenario in models.DamageScenario.objects.filter(
-                expiration_date__lte=now):
+        scenarios = DamageScenario.objects.filter(expiration_date__lte=now)
+        for damage_scenario in scenarios:
 
             logger.info(
                 "Deleting scenario %d (%s), tasks, events and results..." % (
@@ -34,16 +34,6 @@ class Command(BaseCommand):
             SecuredPeriodicTask.objects.filter(
                 name__contains='Scenario (%05d)' % damage_scenario.id).delete()
 
-            for damage_event in damage_scenario.damageevent_set.all():
-                for damage_event_result in (
-                        damage_event.damageeventresult_set.all()):
-                    if damage_event_result.image:
-                        damage_event_result.image.delete()
-                    damage_event_result.delete()
-                if damage_event.result:
-                    damage_event.result.delete()
-
-                damage_event.delete()
             damage_scenario.delete()
 
         logger.info("Finished.")