+++ lizard_damage/tests/test_models.py
@@ -20,9 +20,6 @@ TESTDATA_DIR = os.path.join(settings.BUILDOUT_DIR, 'testdata')
 
 
 class TestDamageScenario(TestCase):
-    def setUp(self):
-        pass
-
     def test_directory_url(self):
         scenario = factories.DamageScenarioFactory.create()
         self.assertEquals(
@@ -41,6 +38,32 @@ class TestDamageScenario(TestCase):
             models.DamageScenario.objects.filter(
                 name="Testscenario").count(),
             1)
+
+    def test_delete_deletes_files(self):
+        scenario = factories.DamageScenarioFactory.create()
+
+        workdir = scenario.workdir
+
+        testfile = os.path.join(workdir, 'test.txt')
+        f = open(testfile, 'w')
+        f.write("Test!")
+        f.close()
+
+        self.assertTrue(os.path.exists(testfile))
+        scenario.delete()
+        self.assertFalse(os.path.exists(testfile))
+
+    def test_delete_deletes_damage_events(self):
+        scenario = factories.DamageScenarioFactory.create()
+        event = factories.DamageEventFactory.create(scenario=scenario)
+
+        event_id = event.id
+
+        self.assertTrue(
+            models.DamageEvent.objects.filter(pk=event_id).exists())
+        scenario.delete()
+        self.assertFalse(
+            models.DamageEvent.objects.filter(pk=event_id).exists())
 
 
 class TestDamageEvent(TestCase):