+++ CHANGES.rst
@@ -2,7 +2,7 @@ Changelog of lizard-damage
 ===================================================
 
 
-2.0.2 (unreleased)
+2.0.2 (2015-03-26)
 ------------------
 
 - Update cleanup command.